Refactor actions in separate file
