+++ stream-video-android-ui-compose/src/main/kotlin/io/getstream/video/android/compose/ui/components/call/renderer/ParticipantVideo.kt
@@ -25,32 +25,17 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Button
-import androidx.compose.material.Icon
 import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Cancel
-import androidx.compose.material.icons.filled.PushPin
-import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -62,37 +47,32 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.BottomEnd
 import androidx.compose.ui.Alignment.Companion.BottomStart
 import androidx.compose.ui.Alignment.Companion.Center
-import androidx.compose.ui.Alignment.Companion.CenterHorizontally
 import androidx.compose.ui.Alignment.Companion.CenterVertically
-import androidx.compose.ui.Alignment.Companion.TopCenter
+import androidx.compose.ui.Alignment.Companion.TopStart
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalInspectionMode
-import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.lifecycleScope
 import io.getstream.video.android.compose.theme.VideoTheme
 import io.getstream.video.android.compose.ui.components.avatar.LocalAvatarPreviewProvider
-import io.getstream.video.android.compose.ui.components.avatar.UserAvatar
 import io.getstream.video.android.compose.ui.components.avatar.UserAvatarBackground
+import io.getstream.video.android.compose.ui.components.call.pinning.ParticipantAction
+import io.getstream.video.android.compose.ui.components.call.pinning.ParticipantActions
+import io.getstream.video.android.compose.ui.components.call.pinning.pinUnpinActions
 import io.getstream.video.android.compose.ui.components.connection.NetworkQualityIndicator
-import io.getstream.video.android.compose.ui.components.indicator.GenericIndicator
 import io.getstream.video.android.compose.ui.components.indicator.SoundIndicator
 import io.getstream.video.android.compose.ui.components.video.VideoRenderer
 import io.getstream.video.android.core.Call
@@ -105,179 +85,7 @@ import io.getstream.video.android.mock.StreamPreviewDataUtils
 import io.getstream.video.android.mock.previewCall
 import io.getstream.video.android.mock.previewParticipantsList
 import io.getstream.video.android.ui.common.R
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import org.openapitools.client.models.OwnCapability
-
-public class ParticipantAction(
-    public val icon: ImageVector? = null,
-    public val label: String,
-    public val condition: (Call, ParticipantState) -> Boolean = { _, _ -> true },
-    public val action: CoroutineScope.(Call, ParticipantState) -> Unit = { _, _ -> },
-)
-
-public val pinUnpinActions: List<ParticipantAction> = listOf(
-    ParticipantAction(
-        icon = Icons.Filled.PushPin,
-        label = "Pin",
-        condition = { call, participantState ->
-            !call.isPinnedParticipant(participantState.sessionId)
-        },
-        action = { call, participantState ->
-            launch {
-                call.state.pin(participantState.sessionId)
-            }
-        },
-    ),
-    ParticipantAction(
-        icon = Icons.Filled.PushPin,
-        label = "Pin for everyone",
-        condition = { call, participantState ->
-            call.hasCapability(OwnCapability.PinForEveryone) && !call.isPinnedParticipant(participantState.sessionId)
-        },
-        action = { call, participantState ->
-            launch {
-                call.state.pin(participantState.sessionId)
-                call.pinForEveryone(call.type, participantState.sessionId)
-            }
-        },
-    ),
-    ParticipantAction(
-        icon = Icons.Filled.Cancel,
-        label = "Unpin for everyone",
-        condition = { call, participantState ->
-            call.hasCapability(OwnCapability.PinForEveryone) && call.isPinnedParticipant(participantState.sessionId)
-        },
-        action = { call, participantState ->
-            launch {
-                call.state.unpin(participantState.sessionId)
-                call.unpinForEveryone(call.type, participantState.sessionId)
-            }
-        },
-    ),
-    ParticipantAction(
-        icon = Icons.Filled.Cancel,
-        label = "Unpin",
-        condition = { call, participantState ->
-            call.isPinnedParticipant(participantState.sessionId)
-        },
-        action = { call, participantState ->
-            launch {
-                call.state.unpin(participantState.sessionId)
-            }
-        },
-    ),
-)
-
-@Composable
-internal fun BoxScope.ParticipantActions(
-    modifier: Modifier = Modifier,
-    actions: List<ParticipantAction>,
-    call: Call,
-    participant: ParticipantState,
-) {
-    var showDialog by remember {
-        mutableStateOf(false)
-    }
-    if (actions.any {
-            it.condition.invoke(call, participant)
-        }
-    ) {
-        GenericIndicator(
-            modifier = modifier.clickable {
-                showDialog = !showDialog
-            },
-        ) {
-            Icon(
-                imageVector = Icons.Outlined.Settings,
-                contentDescription = "Call actions",
-                tint = Color.White,
-            )
-        }
-
-        if (showDialog) {
-            ParticipantActionsDialog(
-                call = call,
-                participant = participant,
-                actions = actions,
-                onDismiss = {
-                    showDialog = false
-                },
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-internal fun BoxScope.ParticipantActionsDialog(
-    call: Call,
-    participant: ParticipantState,
-    actions: List<ParticipantAction>,
-    onDismiss: () -> Unit = {},
-) {
-    val coroutineScope = LocalLifecycleOwner.current.lifecycleScope
-    val userName by participant.userNameOrId.collectAsStateWithLifecycle()
-    val userImage by participant.image.collectAsStateWithLifecycle()
-    val name = remember {
-        participant.name.value
-    }
-    Dialog(onDismiss) {
-        Column(
-            Modifier
-                .background(VideoTheme.colors.appBackground)
-                .align(Center)
-                .padding(16.dp),
-        ) {
-            UserAvatar(
-                modifier = Modifier
-                    .size(82.dp)
-                    .align(CenterHorizontally)
-                    .aspectRatio(1f),
-                userName = userName,
-                userImage = userImage,
-            )
-            Text(
-                modifier = Modifier.fillMaxWidth(),
-                text = name,
-                color = VideoTheme.colors.textHighEmphasis,
-                textAlign = TextAlign.Center,
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-            FlowRow(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.Center,
-                verticalArrangement = Arrangement.Center,
-            ) {
-                actions.forEach {
-                    if (it.condition.invoke(call, participant)) {
-                        Button(
-                            modifier = Modifier.padding(8.dp),
-                            onClick = {
-                                it.action.invoke(coroutineScope, call, participant)
-                            },
-                        ) {
-                            it.icon?.let { hasIcon ->
-                                Icon(
-                                    imageVector = hasIcon,
-                                    contentDescription = it.label,
-                                    tint = VideoTheme.colors.textHighEmphasis,
-                                )
-                            }
-                            Text(
-                                textAlign = TextAlign.Start,
-                                text = it.label,
-                                color = VideoTheme.colors.textHighEmphasis,
-                            )
-                        }
-                    }
-                }
-            }
-            Spacer(modifier = Modifier.height(16.dp))
-        }
-    }
-}
 
 /**
  * Renders a single participant with a given call, which contains all the call states.
@@ -328,8 +136,8 @@ public fun ParticipantVideo(
     ) -> Unit = { actions, call, participant ->
         ParticipantActions(
             Modifier
-                .align(TopCenter)
-                .padding(top = 8.dp),
+                .align(TopStart)
+                .padding(8.dp),
             actions,
             call,
             participant,
